Desktop/requirements.txt
Flask>=3.0.0
Flask-SQLAlchemy>=3.1.0
pandas>=2.0.0
numpy>=1.24.0
gunicorn>=21.0.0
flask-cors>=4.0.0
python-dateutil>=2.8.0
psycopg2-binary>=2.9.0
PyJWT>=2.8.0
Flask-Limiter>=3.5.0
